workloads/agent/requirements.txt
# Unified agent workload requirements.
# Covers: langchain, langgraph, autogen, haystack (RAG), swe_agent.
# Pins are loosened from the per-framework files to let pip resolve a single
# consistent environment. Tighten only if a specific version is required for
# trace reproducibility.

# --- orchestration / planner frameworks ---
langchain-core>=0.3,<1.3
langgraph>=1.1,<2.0
autogen-agentchat>=0.7,<1.0
autogen-ext[openai]>=0.7,<1.0

# --- haystack-style RAG pipeline ---
# Using looser pins than the original haystack workload so the resolver can
# pick versions compatible with torch brought in by sentence-transformers and
# with numpy brought in by faiss.
sentence-transformers>=2.7,<3.0
faiss-cpu>=1.7,<2.0
numpy>=1.26,<2.0
